+++ web/app/routes/pages.js
@@ -1,0 +1,15 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  titleToken: 'Pages',
+
+  queryParams: {
+    q: {
+      refreshModel: true
+    }
+  },
+
+  model: function (params) {
+    return this.store.find('page', params);
+  },
+});